Deduplicate enabled tables filtering
